% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/width.R
\name{defineGlobal}
\alias{defineGlobal}
\title{Global}
\usage{
defineGlobal(width = 991L)
}
\arguments{
\item{width}{Customise breakpoint of flex layout.}
}
\description{
Define width when layout should expand or collapse.
}
