% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{flexlayout}
\alias{flexlayout}
\title{Layout}
\usage{
flexlayout(
  left,
  center,
  right = NULL,
  left_icon = shiny::icon("filter"),
  right_icon = shiny::icon("info")
)
}
\arguments{
\item{left, center, right}{Content of the page.}

\item{left_icon, right_icon}{Icon to show on the left and right side of the page.}
}
\description{
Flexible layout for Shiny applications.
It splits the page in three columns: left, center, and right.
On tablet and mobile the left and right columns are collapsed
into offcanvas that can be brought to view by clicking on icons.
}
