% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buttons.R
\name{offcanvas_button}
\alias{offcanvas_button}
\title{Button}
\usage{
offcanvas_button(..., class = "")
}
\arguments{
\item{...}{Content of button}

\item{class}{Any additional classes to pass to the button.}
}
\description{
Button to trigger offcanvas.
This is just for convenience you can use any button or clickable element.
}
